requirements.txt
openai
python-dotenv
ipykernel
langchain
chromadb
langchain-community
langchain-openai
sentence-transformers
ipywidgets
gradio
openai_gradio
tqdm
torch
thefuzz